requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
supabase>=2.8.0
pydantic==2.5.3
httpx>=0.27.0
gotrue>=2.9.0
websockets>=15.0

